refactor tests to make them more readable
this splits TestValues_types into multiple test functions focused on
different types of values.  Also, instead of testing multiples types at
a time by having a single large struct with lots of different fields,
this switches to testing lots of small structs with a single field.
This does result in longer code in some cases (specifically slices and
arrays, which tend to span multiple lines), but it makes the individual
cases much easier to read, since the input and expected output are
directly next to each other.

More cleanup to come; this just focused on the TestValues_types func.